@@ -6,31 +6,34 @@
     "dev": "vite dev",
     "build": "vite build",
     "preview": "vite preview",
+    "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+    "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
     "lint": "prettier --plugin-search-dir . --check . && eslint .",
     "format": "prettier --plugin-search-dir . --write .",
     "deploy": "npm run build && firebase deploy"
   },
   "devDependencies": {
-    "@sveltejs/adapter-auto": "^2.1.0",
     "@sveltejs/adapter-cloudflare": "^2.3.1",
-    "@sveltejs/adapter-static": "^2.0.2",
-    "@sveltejs/kit": "^1.20.0",
+    "@sveltejs/kit": "^1.22.3",
     "@types/lodash-es": "^4.17.8",
     "autoprefixer": "^10.4.14",
-    "eslint": "^8.41.0",
-    "eslint-config-prettier": "^8.8.0",
-    "eslint-plugin-svelte3": "^4.0.0",
-    "postcss": "^8.4.24",
-    "prettier": "^2.8.8",
-    "prettier-plugin-svelte": "^2.10.1",
-    "prettier-plugin-tailwindcss": "^0.3.0",
-    "svelte": "^3.59.1",
-    "tailwindcss": "^3.3.2",
-    "vite": "^4.3.9"
+    "eslint": "^8.46.0",
+    "eslint-config-prettier": "^8.9.0",
+    "postcss": "^8.4.27",
+    "prettier": "^3.0.0",
+    "prettier-plugin-svelte": "^3.0.3",
+    "prettier-plugin-tailwindcss": "^0.4.1",
+    "svelte": "^4.1.1",
+    "svelte-check": "^3.4.6",
+    "tailwindcss": "^3.3.3",
+    "tslib": "^2.6.1",
+    "typescript": "^5.1.6",
+    "vite": "^4.4.7"
   },
   "type": "module",
   "dependencies": {
-    "firebase": "^9.22.1",
+    "clsx": "^2.0.0",
+    "firebase": "^10.1.0",
     "lodash-es": "^4.17.21",
     "nanoid": "^4.0.2"
   }
