@@ -1,13 +1,6 @@
 <script lang="ts">
   import clsx from 'clsx'
-  import {
-    doc,
-    getDoc,
-    setDoc,
-    addDoc,
-    collection,
-    serverTimestamp,
-  } from 'firebase/firestore'
+  import { doc, getDoc, setDoc, serverTimestamp } from 'firebase/firestore'
   import { db, user, storage } from '$lib/firebase'
   import Input from '$lib/components/Input.svelte'
   import Select from '$lib/components/Select.svelte'
